+++ src/Akeneo/Pim/Automation/DataQualityInsights/back/Application/GetProductTitleSuggestion.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Akeneo\Pim\Automation\DataQualityInsights\Application;
 
 use Akeneo\Pim\Automation\DataQualityInsights\Application\CriteriaEvaluation\Consistency\Text\EvaluateTitleFormatting;
+use Akeneo\Pim\Automation\DataQualityInsights\Domain\Query\GetDescendantVariantProductIdsQueryInterface;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Query\GetIgnoredProductTitleSuggestionQueryInterface;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\Query\GetLatestCriteriaEvaluationsByProductIdQueryInterface;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\ChannelCode;
@@ -30,15 +31,33 @@ class GetProductTitleSuggestion
 
     private $getIgnoredProductTitleSuggestionQuery;
 
+    private $getDescendantVariantProductIdsQuery;
+
     public function __construct(
         GetLatestCriteriaEvaluationsByProductIdQueryInterface $getLatestCriteriaEvaluationsByProductIdQuery,
-        GetIgnoredProductTitleSuggestionQueryInterface $getIgnoredProductTitleSuggestionQuery
+        GetIgnoredProductTitleSuggestionQueryInterface $getIgnoredProductTitleSuggestionQuery,
+        GetDescendantVariantProductIdsQueryInterface $getDescendantVariantProductIdsQuery
     ) {
         $this->getLatestCriteriaEvaluationsByProductIdQuery = $getLatestCriteriaEvaluationsByProductIdQuery;
         $this->getIgnoredProductTitleSuggestionQuery = $getIgnoredProductTitleSuggestionQuery;
+        $this->getDescendantVariantProductIdsQuery = $getDescendantVariantProductIdsQuery;
     }
 
-    public function get(ProductId $productId, ChannelCode $channel, LocaleCode $locale): ?string
+    public function get(ProductId $productId, ChannelCode $channel, LocaleCode $locale, string $productType): ?string
+    {
+        if ($productType === 'product_model') {
+            $variantProductIds = $this->getDescendantVariantProductIdsQuery->fromProductModelIds([strval($productId)]);
+            if (empty($variantProductIds)) {
+                return null;
+            }
+
+            return $this->getTitleSuggestionForAProduct(new ProductId((int) current($variantProductIds)), $channel, $locale);
+        }
+
+        return $this->getTitleSuggestionForAProduct($productId, $channel, $locale);
+    }
+
+    private function getTitleSuggestionForAProduct(ProductId $productId, ChannelCode $channel, LocaleCode $locale)
     {
         $evaluation = $this->getLatestCriteriaEvaluationsByProductIdQuery->execute($productId);
         $criterionEvaluation = $evaluation->get(new CriterionCode(EvaluateTitleFormatting::CRITERION_CODE));
@@ -61,6 +80,6 @@ class GetProductTitleSuggestion
             return null;
         }
 
-        return  $titleSuggestion;
+        return $titleSuggestion;
     }
 }