+++ src/Akeneo/Pim/Automation/DataQualityInsights/back/Infrastructure/Persistence/Query/CachedGetProductRawValuesQuery.php
@@ -41,7 +41,16 @@ class CachedGetProductRawValuesQuery implements GetProductRawValuesQueryInterfac
         }
 
         $query = <<<SQL
-SELECT raw_values FROM pim_catalog_product WHERE id = :product_id;
+SELECT
+    JSON_MERGE(
+        COALESCE(pm1.raw_values, '{}'),
+        COALESCE(pm2.raw_values, '{}'),
+        product.raw_values
+    ) AS raw_values
+FROM pim_catalog_product as product
+    LEFT JOIN pim_catalog_product_model pm1 ON product.product_model_id = pm1.id
+    LEFT JOIN pim_catalog_product_model pm2 ON pm1.parent_id = pm2.id
+WHERE product.id = :product_id;
 SQL;
 
         $statement = $this->db->executeQuery($query,