+++ src/Akeneo/Pim/Automation/DataQualityInsights/front/src/domain/Product.interface.ts
@@ -13,4 +13,7 @@ interface Meta {
   label: {
     [locale: string]: string;
   };
+  level: null | number;
+  attributes_for_this_level: string[];
+  model_type: "product" | "product_model";
 }