+++ src/Akeneo/Pim/Automation/DataQualityInsights/front/src/infrastructure/hooks/EditorHighlight/SuggestedTitle/useFetchTitleSuggestion.ts
@@ -2,7 +2,7 @@ import {useEffect} from 'react';
 import {useDispatch} from 'react-redux';
 import {WidgetElement} from "../../../../application/helper";
 import {fetchTitleSuggestion} from "../../../fetcher";
-import {useCatalogContext} from "../../index";
+import {useCatalogContext, useProduct} from "../../index";
 import {updateWidgetContentAnalysis} from "../../../reducer";
 import useProductAxesRates from "../../useProductAxesRates";
 
@@ -10,37 +10,40 @@ const useFetchTitleSuggestion = (widget: WidgetElement) => {
   const {locale, channel} = useCatalogContext();
   const {axesRates, productId} = useProductAxesRates();
   const {analysis} = widget;
+  const product = useProduct();
   const dispatchAction = useDispatch();
 
   useEffect(() => {
-    (async () => {
-      if (!widget.isMainLabel) {
-        return;
-      }
+    if (axesRates !== undefined) { // on app initialization, the axes rates data is undefined for the current product
+      (async () => {
+        if (!widget.isMainLabel) {
+          return;
+        }
 
-      if (!productId || !channel || !locale) {
-        dispatchAction(updateWidgetContentAnalysis(widget.id, []));
-        return;
-      }
+        if (!productId || !channel || !locale) {
+          dispatchAction(updateWidgetContentAnalysis(widget.id, []));
+          return;
+        }
 
-      const result: string|null = await fetchTitleSuggestion(productId, channel, locale);
+        const result: string|null = await fetchTitleSuggestion(product, channel, locale);
 
-      if (typeof result !== "string" || result.length === 0) {
-        dispatchAction(updateWidgetContentAnalysis(widget.id, []));
-        return;
-      }
+        if (typeof result !== "string" || result.length === 0) {
+          dispatchAction(updateWidgetContentAnalysis(widget.id, []));
+          return;
+        }
 
-      const suggestions: string[] = [result];
+        const suggestions: string[] = [result];
 
-      dispatchAction(updateWidgetContentAnalysis(widget.id, [{
-        text: widget.content,
-        type: "title_suggestion",
-        globalOffset: 0,
-        offset: 0,
-        line: 1,
-        suggestions
-      }]));
-    })();
+        dispatchAction(updateWidgetContentAnalysis(widget.id, [{
+          text: widget.content,
+          type: "title_suggestion",
+          globalOffset: 0,
+          offset: 0,
+          line: 1,
+          suggestions
+        }]));
+      })();
+    }
   }, [axesRates]);
 
   return {analysis};