+++ src/Akeneo/Pim/Automation/DataQualityInsights/front/src/infrastructure/hooks/useProductAxesRates.ts
@@ -1,19 +1,29 @@
-import {useSelector} from "react-redux";
+import {useEffect} from "react";
+import {useSelector, useDispatch} from "react-redux";
 import {ProductEditFormState} from "../store";
+import {getProductAxesRatesAction} from "../reducer";
 
 
 const useProductAxesRates = () => {
-  const {axesRates, productId, productUpdated} = useSelector((state: ProductEditFormState) => {
+  const dispatchAction = useDispatch();
+
+  const {productId, productUpdated, axesRates} = useSelector((state: ProductEditFormState) => {
     const productId = state.product.meta.id;
     const productUpdated = state.product.updated;
-    const axesRates = productId ? state.productAxesRates[productId] : {};
+    const axesRates = productId ? state.productAxesRates[productId] : undefined;
 
     return {
-      axesRates: axesRates || {},
       productId: productId,
       productUpdated,
+      axesRates
     };
   });
+
+  useEffect(() => {
+    if (productId && axesRates === undefined) {
+      dispatchAction(getProductAxesRatesAction(productId, {}));
+    }
+  }, [productId, axesRates]);
 
   return {
     axesRates,