+++ src/Akeneo/Pim/Automation/DataQualityInsights/front/src/infrastructure/reducer/productReducer.ts
@@ -29,6 +29,9 @@ const initialState: ProductState = {
   meta: {
     id: null,
     label: {},
+    attributes_for_this_level: [],
+    level: null,
+    model_type: "product",
   },
 };