+++ CHANGELOG.md
@@ -7,3 +7,7 @@
 - DAPI-697: Data quality - As Julia, when I'm on the DQI page, I want to click the attributes that need improvements and land on the PEF.
 - DAPI-806: Improve criteria evaluations performance
 - DAPI-739: Add coefficients by criterion to the calculation of the axes rates
+
+## New features
+
+- DAPI-854: Data quality - Variant products are also evaluated